Created the options menu. When the options menu opens, it pauses all other stages. To make this work I've changed how inputs are processed. Each stage has a step function that runs every frame. In this function, the Q.inputs object is evaluated.
Access the options menu by pressing the spacebar (or z). Up and down arrows for movement and enter for changing the values. Press spacebar again to close this menu.

I also added the start of the Action menu for battles. When hovering a Character, press enter to access this menu. Press esc to get rid of it (enter should be reserved for confirming the selected option).