Added the inventory system with the necessary json files. There is now a Bag object that is in charge of adding and removing items from the bag.
Added the menu functions for when you select active character. Was able to get the movement code in using the A* algorithm. You can test this by selecting the player and choosing 'move'. This will load the grid of possible tiles that you can move on. Selecting one of those tiles moves the player to it. Each tile is weighted so you will notice that it takes more 'move' to walk across the sand.