Made the menuControls component which adds functionality to the list style menus. This component is used by the action menu that is used to control characters in battle and also by the big status menu that is selected from the action menu.
Features:

Up/Down - The component checks if the selection is above or below the list and cycles it to the proper place. The menu gets the hoverOption event triggered every time the user presses up/down.

Confirm - When the user presses 'confirm', the menu gets an event that triggers pressConfirm.

Esc - The menu gets a 'pressBack' event triggered.

skipGray - The menuControls can skip 'greyed out' options such as when the user has moved, don't allow for move to be selected.