Onload effects/conditions now work. Disabled/enabled now works. We can start creating some pretty complex events now!
Next on the list: make scene and event variables happen. After that, I'll include a way to 'link' different types of events (battle, and battle scenes), though this won't be useful until there is an editor for those events. At least we'll be able to link multiple story events together as a proof of concept.