Reworked the flavour events a bit.
To determine which event to play, there is a priority for each event. There is also a 'recur' property which says if an event can have a chance to play again if it was out-prioritized by a different event last time. An example of use:

There is a character that has an event triggered after visiting them 5 times in the infirmary.
Non-recurring event -> Thanks for visiting me 5 times now!
Recurring event -> Thanks for visiting so many times!

Characters only save their completed events and the game looks at the master list of events each time an event may be triggered.