Fixed a few small bugs.
Created the save-event-references.php file which takes data when an event is saved and saves the references to events, sceneVars, and globalVars. This is used in the flowchart to know which event should point to which other event. It is also used in the vars editor to display a list of events that point to that var.