Fixed a bunch of things including:
Made base stat modifying techniques properly apply the modification to combatStats. Instead of modifying baseStats directly, there is a modifiable baseStat stored within combatStats.

Added a 'set' operator for evaluateStringConditional. This will allow variables to be checked if they are set (or not). This could be used in not allowing a battle to happen twice, but also saving 'defeat' or 'victory' to the variable that determines if the battle has taken place.

Lots more but I forgot..