Added helper tile for attack and move. This tile follows the location of the mouse.
Move, Attack, and End Turn all work. Technique and Item are almost working. Carry and Status haven't been started.

I'll need to come up with a design for status. It'll show the same information for the current turn character as well as the other characters you can check status during a turn. Optimally this would be the same status menu that is displayed in locations.